Fixed remove all services script

+++ remove-all-services.sh
@@ -1,16 +1,19 @@
 echo "Removing all GrowSense services..."
 
-# TODO: This should be moved to a different script. It doesn't just remove device service scripts but removes all GrowSense services.
 for filename in /lib/systemd/system/growsense-*.service; do
-  [[ ! -f "$filename" ]] || break
-  shortname=$(basename $filename)
-  echo "Removing service: $shortname"
-  echo ""
-  sh $SYSTEMCTL_SCRIPT stop "$shortname" || echo "Failed to stop $shortname service. Skipping."
-  sh $SYSTEMCTL_SCRIPT disable "$shortname" || echo "Failed to disable $shortname service. Skipping."
+  if [[ -f "$filename" ]]; then
+    shortname=$(basename $filename)
+    echo "Removing service: $shortname"
+    echo ""
+
+    bash systemctl.sh stop "$shortname" || echo "Failed to stop $shortname service. Skipping."
+    bash systemctl.sh disable "$shortname" || echo "Failed to disable $shortname service. Skipping."
   
-  if [ ! -f "is-mock-systemctl.txt" ]; then
-    $SUDO rm -v $filename || echo "Failed to remove $shortname service. Skipping."
+    if [ ! -f "is-mock-systemctl.txt" ]; then
+      $SUDO rm -v $filename || echo "Failed to remove $shortname service. Skipping."
+    else
+      echo "[mock] $SUDO rm -v $filename"
+    fi
   fi
   echo "" 
 done
